+++ truck-stepio/src/in/step_geometry/geom_impls.rs
@@ -67,8 +67,7 @@ impl ToSameGeometry<Surface> for RevolutedCurve<Curve3D> {
     fn to_same_geometry(&self) -> Surface {
         let default = || {
             let (curve, origin, axis) = (self.entity_curve().inverse(), self.origin(), self.axis());
-            let mut processor = Processor::new(RevolutedCurve::by_revolution(curve, origin, axis));
-            processor.invert();
+            let processor = Processor::new(RevolutedCurve::by_revolution(curve, origin, axis));
             Surface::SweptCurve(SweptCurve::RevolutedCurve(processor))
         };
         match self.entity_curve() {
@@ -78,11 +77,9 @@ impl ToSameGeometry<Surface> for RevolutedCurve<Curve3D> {
                 let axis = self.axis();
                 if v.cross(axis).so_small() {
                     let o = self.origin();
-                    let origin = o + (q - o).dot(axis) * axis;
-                    let line = Line(q, q - v.normalize());
-                    let revo = RevolutedCurve::by_revolution(line, origin, axis);
-                    let mut processor = Processor::new(revo);
-                    processor.invert();
+                    let origin = o + (p - o).dot(axis) * axis;
+                    let revo = RevolutedCurve::by_revolution(*line, origin, axis);
+                    let processor = Processor::new(revo);
                     Surface::ElementarySurface(ElementarySurface::CylindricalSurface(processor))
                 } else {
                     default()